Add DispatchingStoreType protocol
Defines the interface of a dispatching, stateless Store. Useful for dependency injection when your type only cares about dispatching raw actions, and doesn’t need to know about the store’s state.

Fixes #174
